funciona

+++ src/main/java/ec/edu/ups/icc/fundamentos01/users/dtos/UserResponseDto.java
@@ -1,10 +1,7 @@
 package ec.edu.ups.icc.fundamentos01.users.dtos;
 
-/**
- * DTO de respuesta para usuarios
- */
 public class UserResponseDto {
-    public Long id;
+    public int id;
     public String name;
     public String email;
 }
